+++ src/config.lua
@@ -4,24 +4,23 @@ local config = {
 	AutoHarvestOnExit = {
 		Enabled = false,
 	},
+	BypassToolRequirements = {
+		Enabled = false
+	},
 	Harvest = {
 		Enabled = true,
 	},
 	Shovel = {
 		Enabled = true,
-		GiveTool = false
 	},
 	Pickaxe = {
 		Enabled = true,
-		GiveTool = false
 	},
 	Exorcism = {
 		Enabled = true,
-		GiveTool = false
 	},
 	Fishing = {
 		Enabled = true,
-		GiveTool = false
 	},
 	logLevel = 0,
 }
@@ -31,24 +30,23 @@ local description = {
 	AutoHarvestOnExit = {
 		Enabled = "Auto Harvest everything on room exit, without the need to interact with each point."
 	},
+	BypassToolRequirements = {
+		Enabled = "Bypass Tool requirements for harvesting."
+	},
 	Harvest = {
 		Enabled = "Enable to skip harvest animations."
 	},
 	Shovel = {
 		Enabled = "Enable to skip shovel animations.",
-		GiveTool = "Sets to true if you want the tool, but haven`t."
 	},
 	Pickaxe = {
 		Enabled = "Enable to skip pickaxe animations.",
-		GiveTool = "Sets to true if you want the tool, but haven`t."
 	},
 	Exorcism = {
 		Enabled = "Enable to skip exorcism animations.",
-		GiveTool = "Sets to true if you want the tool, but haven`t."
 	},
 	Fishing = {
 		Enabled = "Enable to skip fishing animations.",
-		GiveTool = "Sets to true if you want the tool, but haven`t."
 	},
 
 	logLevel = "Log Level for prints in console."