+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added _AutoHarvestOnExit_. Harvest everything on room exit, without the need to interact with each point.
+
 ## [0.5.0] - 2025-10-09
 
 ### Modified